# api/swagger/swagger.yaml
basePath: /api
info:
  contact:
    email: abdoelrachmad@gmail.com
    name: Maintainer
  description: Simple blog server built with go.
  title: GoBlog Server
  version: "1.0"
paths:
...
  /v1/auth/editor/comment/{uid}:
    delete:
      description: Delete a comment (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete Comment (Soft)
      tags:
      - Comment (Editor)
    get:
      description: Get a comment.
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      content:
                        type: string
                      createdAt:
                        type: string
                      deletedAt:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      parentCommentUid:
                        type: string
                      postUid:
                        type: string
                      replyCount:
                        type: integer
                      uid:
                        type: string
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Comment
      tags:
      - Comment (Editor)
  /v1/auth/editor/comment/{uid}/detrash:
    patch:
      description: Restore deleted comment (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore Comment (Soft)
      tags:
      - Comment (Editor)
    put:
      description: Restore deleted comment (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore Comment (Soft)
      tags:
      - Comment (Editor)
  /v1/auth/editor/comment/{uid}/permanent:
    delete:
      description: Delete comment (permanent).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete Comment (Permanent)
      tags:
      - Comment (Editor)
  /v1/auth/editor/comments:
    get:
      description: Get comments.
      parameters:
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=published, ?type=draft.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        content:
                          type: string
                        createdAt:
                          type: string
                        deletedAt:
                          type: string
                        email:
                          type: string
                        name:
                          type: string
                        parentCommentUid:
                          type: string
                        postUid:
                          type: string
                        replyCount:
                          type: integer
                        uid:
                          type: string
                      type: object
                  type: array
              type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Comments
      tags:
      - Comment (Editor)
  /v1/auth/editor/comments/stats:
    get:
      description: Get comments's stats.
      parameters:
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=published, ?type=draft.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      currentPage:
                        type: integer
                      itemsPerPage:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Comments Stats
      tags:
      - Comment (Editor)
...
  /v1/auth/editor/post/{uid}/comments:
    get:
      description: Get post's comments.
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=active.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        content:
                          type: string
                        createdAt:
                          type: string
                        deletedAt:
                          type: string
                        email:
                          type: string
                        name:
                          type: string
                        parentCommentUid:
                          type: string
                        postUid:
                          type: string
                        replyCount:
                          type: integer
                        uid:
                          type: string
                      type: object
                  type: array
              type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Post Comments
      tags:
      - Comment (Editor)
  /v1/auth/editor/post/{uid}/comments/stats:
    get:
      description: Get post's comments stats.
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=active.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      currentPage:
                        type: integer
                      itemsPerPage:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Post Comments Stats
      tags:
      - Comment (Editor)
...
securityDefinitions:
  BearerAuth:
    description: '`Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"